Third attempt at long to int changes.
Done with:
    sed -i -e 's/unsigned long/uint32_t/g' *.cpp *.h
    sed -i -e 's/ulong/uint32_t/g' *.cpp *.h
    sed -i -e 's/signed long/int32_t/g' *.cpp *.h
    sed -i -e 's/long/int32_t/g' *.cpp *.h
    sed -i -e 's/uint32_t int32_t/unsigned long long/g' *.cpp *.h
    sed -i -e 's/int32_t int32_t/long long/g' *.cpp *.h

+++ src/Area.cpp
@@ -13,7 +13,7 @@
 namespace Areas
 {
 	
-	Area SetArea(long left, long top, long right, long bottom) {
+	Area SetArea(int32_t left, int32_t top, int32_t right, int32_t bottom) {
 		Area a;
 
 		a.left = left;
@@ -23,7 +23,7 @@ namespace Areas
 		return a;
 	} // SetArea
 
-	Area SetMarquee(long left, long top, long right, long bottom) {
+	Area SetMarquee(int32_t left, int32_t top, int32_t right, int32_t bottom) {
 		Area a;
 
 		a.left = Min(left, right);
@@ -33,17 +33,17 @@ namespace Areas
 		return a;
 	} // SetMarquee
 
-	void OffsetArea(Area *a, long dx, long dy) {
+	void OffsetArea(Area *a, int32_t dx, int32_t dy) {
 		a->left += dx; a->right += dx;
 		a->top += dy; a->bottom += dy;
 	} // OffsetArea
 
-	void ScaleArea(Area *a, long scale) {
+	void ScaleArea(Area *a, int32_t scale) {
 		a->left *= scale; a->right *= scale;
 		a->top *= scale; a->bottom *= scale;
 	} // ScaleArea
 
-	Area InsetArea(const Area *area, long dx, long dy) {
+	Area InsetArea(const Area *area, int32_t dx, int32_t dy) {
 		Area inset;
 
 		inset.left = area->left + dx; inset.right = Max(inset.left, area->right - dx);
@@ -76,8 +76,8 @@ namespace Areas
 		return c;
 	} // IntersectArea
 
-	long SubtractArea(const Area *a, const Area *b, Area part[]) { // returns #parts (0..4)
-		long left, top, right, bottom, parts;
+	int32_t SubtractArea(const Area *a, const Area *b, Area part[]) { // returns #parts (0..4)
+		int32_t left, top, right, bottom, parts;
 
 		left = a->left;  top = a->top;
 		right = a->right; bottom = a->bottom;
@@ -123,7 +123,7 @@ namespace Areas
 		return a->left == b->left && a->top == b->top && a->right == b->right && a->bottom == b->bottom;
 	} // SameArea
 
-	bool AreaIncludesPoint(const Area *a, long x, long y) {
+	bool AreaIncludesPoint(const Area *a, int32_t x, int32_t y) {
 		return a->left <= x && x <= a->right && a->top <= y && y <= a->bottom;
 	} // AreaIncludesPoint
